Add post images
