--- tsconfig-codegen.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2018",
-    "moduleResolution": "node16",
-    "module": "node16",
-    "outDir": "dist"
-  },
-  "include": ["src/**/*"]
-}
